wip: move to generate config per component
