--- cmd/configschema/README.md
@@ -1,5 +1,0 @@
-# ConfigSchema API
-
-This package contains an API that can be used to introspect the configuration
-struct of a collector component. It can be used to generate documentation or
-tools to help users configure the collector.
