--- cmd/configschema/docsgen/docsgen/testdata/test.tmpl
@@ -1,1 +1,0 @@
-{{ .Type }}
