--- internal/components/components.go
@@ -1,413 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package components // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/components"
-
-import (
-	"go.opentelemetry.io/collector/connector"
-	"go.opentelemetry.io/collector/connector/forwardconnector"
-	"go.opentelemetry.io/collector/exporter"
-	"go.opentelemetry.io/collector/exporter/debugexporter"
-	"go.opentelemetry.io/collector/exporter/loggingexporter"
-	"go.opentelemetry.io/collector/exporter/otlpexporter"
-	"go.opentelemetry.io/collector/exporter/otlphttpexporter"
-	"go.opentelemetry.io/collector/extension"
-	"go.opentelemetry.io/collector/extension/ballastextension"
-	"go.opentelemetry.io/collector/extension/zpagesextension"
-	"go.opentelemetry.io/collector/otelcol"
-	"go.opentelemetry.io/collector/processor"
-	"go.opentelemetry.io/collector/processor/batchprocessor"
-	"go.opentelemetry.io/collector/processor/memorylimiterprocessor"
-	"go.opentelemetry.io/collector/receiver"
-	"go.opentelemetry.io/collector/receiver/otlpreceiver"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/countconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/datadogconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/exceptionsconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/routingconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/servicegraphconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/connector/spanmetricsconnector"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/alibabacloudlogserviceexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awscloudwatchlogsexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awsemfexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awskinesisexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/awsxrayexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/azuredataexplorerexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/azuremonitorexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/carbonexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/cassandraexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/clickhouseexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/coralogixexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datasetexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/dynatraceexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/f5cloudexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/fileexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/googlecloudexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/googlecloudpubsubexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/googlemanagedprometheusexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/influxdbexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/instanaexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/kafkaexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/loadbalancingexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/logicmonitorexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/logzioexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/lokiexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/mezmoexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/opencensusexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/opensearchexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/pulsarexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sapmexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sentryexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/skywalkingexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/sumologicexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/tanzuobservabilityexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/tencentcloudlogserviceexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/zipkinexporter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/asapauthextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/awsproxy"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/basicauthextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/bearertokenauthextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/headerssetterextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/healthcheckextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/httpforwarder"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/jaegerremotesampling"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/oauth2clientauthextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/dockerobserver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/ecstaskobserver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/hostobserver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/observer/k8sobserver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/oidcauthextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/pprofextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/sigv4authextension"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/storage/dbstorage"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/extension/storage/filestorage"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/attributesprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/cumulativetodeltaprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/datadogprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatorateprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/filterprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/groupbyattrsprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/groupbytraceprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/metricsgenerationprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/metricstransformprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/probabilisticsamplerprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourcedetectionprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourceprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/routingprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/servicegraphprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/spanmetricsprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/spanprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/tailsamplingprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/transformprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/activedirectorydsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/aerospikereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/apachereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/apachesparkreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awscloudwatchreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awscontainerinsightreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awsecscontainermetricsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awsfirehosereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/awsxrayreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/azureblobreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/azureeventhubreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/azuremonitorreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/bigipreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/carbonreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/chronyreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/cloudfoundryreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/collectdreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/couchdbreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/datadogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/dockerstatsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/elasticsearchreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/expvarreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/filelogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/flinkmetricsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/fluentforwardreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/googlecloudpubsubreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/googlecloudspannerreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/haproxyreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/hostmetricsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/httpcheckreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/iisreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/influxdbreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/jaegerreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/jmxreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/journaldreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8sclusterreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8seventsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/k8sobjectsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/kafkametricsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/kafkareceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/kubeletstatsreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/lokireceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/memcachedreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/mongodbatlasreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/mongodbreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/mysqlreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/nginxreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/nsxtreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/opencensusreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/oracledbreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/otlpjsonfilereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/podmanreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/postgresqlreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/prometheusreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/pulsarreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/purefareceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/purefbreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/rabbitmqreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/receivercreator"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/redisreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/riakreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/saphanareceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/sapmreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/signalfxreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/simpleprometheusreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/skywalkingreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/snmpreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/snowflakereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/solacereceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/sqlqueryreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/sqlserverreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/sshcheckreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/statsdreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/syslogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/tcplogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/udplogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/vcenterreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/wavefrontreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/webhookeventreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowseventlogreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowsperfcountersreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/zipkinreceiver"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/zookeeperreceiver"
-)
-
-func Components() (otelcol.Factories, error) {
-	var err error
-	factories := otelcol.Factories{}
-	extensions := []extension.Factory{
-		asapauthextension.NewFactory(),
-		awsproxy.NewFactory(),
-		ballastextension.NewFactory(),
-		basicauthextension.NewFactory(),
-		bearertokenauthextension.NewFactory(),
-		dbstorage.NewFactory(),
-		dockerobserver.NewFactory(),
-		ecstaskobserver.NewFactory(),
-		filestorage.NewFactory(),
-		headerssetterextension.NewFactory(),
-		healthcheckextension.NewFactory(),
-		hostobserver.NewFactory(),
-		httpforwarder.NewFactory(),
-		jaegerremotesampling.NewFactory(),
-		k8sobserver.NewFactory(),
-		pprofextension.NewFactory(),
-		oauth2clientauthextension.NewFactory(),
-		oidcauthextension.NewFactory(),
-		sigv4authextension.NewFactory(),
-		zpagesextension.NewFactory(),
-	}
-	factories.Extensions, err = extension.MakeFactoryMap(extensions...)
-	if err != nil {
-		return otelcol.Factories{}, err
-	}
-
-	receivers := []receiver.Factory{
-		activedirectorydsreceiver.NewFactory(),
-		aerospikereceiver.NewFactory(),
-		apachereceiver.NewFactory(),
-		apachesparkreceiver.NewFactory(),
-		awscontainerinsightreceiver.NewFactory(),
-		awsecscontainermetricsreceiver.NewFactory(),
-		awsfirehosereceiver.NewFactory(),
-		awscloudwatchreceiver.NewFactory(),
-		awsxrayreceiver.NewFactory(),
-		azureblobreceiver.NewFactory(),
-		azureeventhubreceiver.NewFactory(),
-		azuremonitorreceiver.NewFactory(),
-		bigipreceiver.NewFactory(),
-		carbonreceiver.NewFactory(),
-		chronyreceiver.NewFactory(),
-		cloudfoundryreceiver.NewFactory(),
-		collectdreceiver.NewFactory(),
-		couchdbreceiver.NewFactory(),
-		datadogreceiver.NewFactory(),
-		dockerstatsreceiver.NewFactory(),
-		elasticsearchreceiver.NewFactory(),
-		expvarreceiver.NewFactory(),
-		filelogreceiver.NewFactory(),
-		flinkmetricsreceiver.NewFactory(),
-		fluentforwardreceiver.NewFactory(),
-		googlecloudspannerreceiver.NewFactory(),
-		googlecloudpubsubreceiver.NewFactory(),
-		haproxyreceiver.NewFactory(),
-		hostmetricsreceiver.NewFactory(),
-		httpcheckreceiver.NewFactory(),
-		influxdbreceiver.NewFactory(),
-		iisreceiver.NewFactory(),
-		jaegerreceiver.NewFactory(),
-		jmxreceiver.NewFactory(),
-		journaldreceiver.NewFactory(),
-		kafkareceiver.NewFactory(),
-		kafkametricsreceiver.NewFactory(),
-		k8sclusterreceiver.NewFactory(),
-		k8seventsreceiver.NewFactory(),
-		k8sobjectsreceiver.NewFactory(),
-		kubeletstatsreceiver.NewFactory(),
-		lokireceiver.NewFactory(),
-		memcachedreceiver.NewFactory(),
-		mongodbatlasreceiver.NewFactory(),
-		mongodbreceiver.NewFactory(),
-		mysqlreceiver.NewFactory(),
-		nsxtreceiver.NewFactory(),
-		nginxreceiver.NewFactory(),
-		opencensusreceiver.NewFactory(),
-		oracledbreceiver.NewFactory(),
-		otlpjsonfilereceiver.NewFactory(),
-		otlpreceiver.NewFactory(),
-		podmanreceiver.NewFactory(),
-		postgresqlreceiver.NewFactory(),
-		prometheusreceiver.NewFactory(),
-		pulsarreceiver.NewFactory(),
-		purefareceiver.NewFactory(),
-		purefbreceiver.NewFactory(),
-		rabbitmqreceiver.NewFactory(),
-		receivercreator.NewFactory(),
-		redisreceiver.NewFactory(),
-		riakreceiver.NewFactory(),
-		saphanareceiver.NewFactory(),
-		sapmreceiver.NewFactory(),
-		signalfxreceiver.NewFactory(),
-		simpleprometheusreceiver.NewFactory(),
-		skywalkingreceiver.NewFactory(),
-		snmpreceiver.NewFactory(),
-		snowflakereceiver.NewFactory(),
-		solacereceiver.NewFactory(),
-		splunkhecreceiver.NewFactory(),
-		sqlqueryreceiver.NewFactory(),
-		sqlserverreceiver.NewFactory(),
-		sshcheckreceiver.NewFactory(),
-		statsdreceiver.NewFactory(),
-		wavefrontreceiver.NewFactory(),
-		webhookeventreceiver.NewFactory(),
-		windowseventlogreceiver.NewFactory(),
-		windowsperfcountersreceiver.NewFactory(),
-		zookeeperreceiver.NewFactory(),
-		syslogreceiver.NewFactory(),
-		tcplogreceiver.NewFactory(),
-		udplogreceiver.NewFactory(),
-		vcenterreceiver.NewFactory(),
-		zipkinreceiver.NewFactory(),
-	}
-	factories.Receivers, err = receiver.MakeFactoryMap(receivers...)
-	if err != nil {
-		return otelcol.Factories{}, err
-	}
-
-	exporters := []exporter.Factory{
-		alibabacloudlogserviceexporter.NewFactory(),
-		awscloudwatchlogsexporter.NewFactory(),
-		awsemfexporter.NewFactory(),
-		awskinesisexporter.NewFactory(),
-		awsxrayexporter.NewFactory(),
-		azuredataexplorerexporter.NewFactory(),
-		azuremonitorexporter.NewFactory(),
-		carbonexporter.NewFactory(),
-		cassandraexporter.NewFactory(),
-		clickhouseexporter.NewFactory(),
-		coralogixexporter.NewFactory(),
-		datadogexporter.NewFactory(),
-		datasetexporter.NewFactory(),
-		debugexporter.NewFactory(),
-		dynatraceexporter.NewFactory(),
-		elasticsearchexporter.NewFactory(),
-		f5cloudexporter.NewFactory(),
-		fileexporter.NewFactory(),
-		googlecloudexporter.NewFactory(),
-		googlemanagedprometheusexporter.NewFactory(),
-		googlecloudpubsubexporter.NewFactory(),
-		influxdbexporter.NewFactory(),
-		instanaexporter.NewFactory(),
-		kafkaexporter.NewFactory(),
-		loadbalancingexporter.NewFactory(),
-		loggingexporter.NewFactory(),
-		logicmonitorexporter.NewFactory(),
-		logzioexporter.NewFactory(),
-		lokiexporter.NewFactory(),
-		mezmoexporter.NewFactory(),
-		opencensusexporter.NewFactory(),
-		opensearchexporter.NewFactory(),
-		otlpexporter.NewFactory(),
-		otlphttpexporter.NewFactory(),
-		prometheusexporter.NewFactory(),
-		prometheusremotewriteexporter.NewFactory(),
-		pulsarexporter.NewFactory(),
-		sapmexporter.NewFactory(),
-		sentryexporter.NewFactory(),
-		signalfxexporter.NewFactory(),
-		skywalkingexporter.NewFactory(),
-		splunkhecexporter.NewFactory(),
-		sumologicexporter.NewFactory(),
-		tanzuobservabilityexporter.NewFactory(),
-		tencentcloudlogserviceexporter.NewFactory(),
-		zipkinexporter.NewFactory(),
-	}
-	factories.Exporters, err = exporter.MakeFactoryMap(exporters...)
-	if err != nil {
-		return otelcol.Factories{}, err
-	}
-
-	processors := []processor.Factory{
-		attributesprocessor.NewFactory(),
-		batchprocessor.NewFactory(),
-		filterprocessor.NewFactory(),
-		groupbyattrsprocessor.NewFactory(),
-		groupbytraceprocessor.NewFactory(),
-		k8sattributesprocessor.NewFactory(),
-		memorylimiterprocessor.NewFactory(),
-		metricstransformprocessor.NewFactory(),
-		metricsgenerationprocessor.NewFactory(),
-		probabilisticsamplerprocessor.NewFactory(),
-		resourcedetectionprocessor.NewFactory(),
-		resourceprocessor.NewFactory(),
-		routingprocessor.NewFactory(),
-		tailsamplingprocessor.NewFactory(),
-		servicegraphprocessor.NewFactory(),
-		spanmetricsprocessor.NewFactory(),
-		spanprocessor.NewFactory(),
-		cumulativetodeltaprocessor.NewFactory(),
-		datadogprocessor.NewFactory(),
-		deltatorateprocessor.NewFactory(),
-		transformprocessor.NewFactory(),
-	}
-	factories.Processors, err = processor.MakeFactoryMap(processors...)
-	if err != nil {
-		return otelcol.Factories{}, err
-	}
-
-	connectors := []connector.Factory{
-		forwardconnector.NewFactory(),
-		countconnector.NewFactory(),
-		datadogconnector.NewFactory(),
-		exceptionsconnector.NewFactory(),
-		routingconnector.NewFactory(),
-		servicegraphconnector.NewFactory(),
-		spanmetricsconnector.NewFactory(),
-	}
-	factories.Connectors, err = connector.MakeFactoryMap(connectors...)
-	if err != nil {
-		return otelcol.Factories{}, err
-	}
-
-	return factories, nil
-}
