+++ internal/configschema/comments_test.go
@@ -9,8 +9,6 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/collector/config/configtls"
 )
 
 func TestFieldComments(t *testing.T) {
@@ -32,11 +30,4 @@ func TestExternalType(t *testing.T) {
 		"A UUID is a 128 bit (16 byte) Universal Unique IDentifier as defined in RFC\n4122.\n",
 		comments["_struct"],
 	)
-}
-
-func TestSubPackage(t *testing.T) {
-	s := configtls.TLSClientSetting{}
-	v := reflect.ValueOf(s)
-	_, err := commentsForStruct(v, testDR())
-	require.NoError(t, err)
 }
