+++ internal/configschema/doc.go
@@ -1,0 +1,7 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package configschema contains an API that can be used to introspect the configuration
+//struct of a collector component. It can be used to generate documentation or
+//tools to help users configure the collector.
+package configschema
