+++ internal/configschema/docs.go
@@ -1,0 +1,68 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package configschema // import "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema/docsgen/docsgen"
+
+import (
+	"os"
+	"path/filepath"
+	"reflect"
+	"strings"
+	"text/template"
+
+	"go.opentelemetry.io/collector/component"
+)
+
+const mdFileName = "config.md"
+
+func GenerateConfigDoc(srcRoot string,
+	f component.Factory,
+) error {
+	tableTmpl, err := tableTemplate()
+	if err != nil {
+		return err
+	}
+	cfg, err := GetCfgInfo(f)
+	if err != nil {
+		return err
+	}
+
+	dr := NewDirResolver(srcRoot, string(f.Type()))
+	return writeConfigDoc(tableTmpl, dr, cfg, os.WriteFile)
+}
+
+type writeFileFunc func(filename string, data []byte, perm os.FileMode) error
+
+func writeConfigDoc(
+	tableTmpl *template.Template,
+	dr DirResolver,
+	ci CfgInfo,
+	writeFile writeFileFunc,
+) error {
+	v := reflect.ValueOf(ci.CfgInstance)
+	f, err := ReadFields(v, dr)
+	if err != nil {
+		panic(err)
+	}
+
+	mdBytes := renderHeader(string(ci.Type), f.Doc)
+
+	f.Name = typeToName(f.Type)
+
+	tableBytes, err := renderTable(tableTmpl, f)
+	if err != nil {
+		panic(err)
+	}
+	mdBytes = append(mdBytes, tableBytes...)
+
+	if hasTimeDuration(f) {
+		mdBytes = append(mdBytes, durationBlock...)
+	}
+
+	return writeFile(filepath.Join(dr.SrcRoot, mdFileName), mdBytes, 0644)
+}
+
+func typeToName(typ string) string {
+	idx := strings.IndexRune(typ, '.')
+	return typ[:idx]
+}
