+++ internal/configschema/metadata_writer_test.go
@@ -4,7 +4,7 @@
 // skipping windows to avoid this golang bug: https://github.com/golang/go/issues/51442
 //go:build !windows
 
-package cfgmetadatagen
+package configschema
 
 import (
 	"io"
@@ -14,16 +14,14 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema"
 )
 
 func TestMetadataFileWriter(t *testing.T) {
 	tempDir := t.TempDir()
 	w := newMetadataFileWriter(tempDir)
-	err := w.write(configschema.CfgInfo{Group: "mygroup", Type: "mytype"}, []byte("hello"))
+	err := w.write(CfgInfo{Type: "mytype"}, []byte("hello"))
 	require.NoError(t, err)
-	file, err := os.Open(filepath.Join(tempDir, "mygroup", "mytype.yaml"))
+	file, err := os.Open(filepath.Join(tempDir, "mytype.yaml"))
 	require.NoError(t, err)
 	bytes, err := io.ReadAll(file)
 	require.NoError(t, err)
