+++ internal/configschema/render.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package docsgen // import "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema/docsgen/docsgen"
+package configschema // import "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema/docsgen/docsgen"
 
 import (
 	"bytes"
@@ -10,21 +10,18 @@ import (
 
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/configschema"
 )
 
-func renderHeader(typ, group, doc string) []byte {
+func renderHeader(typ, doc string) []byte {
 	caser := cases.Title(language.English)
 	return []byte(fmt.Sprintf(
-		"# %s %s Reference\n\n%s\n\n",
+		"# %s Reference\n\n%s\n\n",
 		caser.String(typ),
-		caser.String(group),
 		doc,
 	))
 }
 
-func renderTable(tmpl *template.Template, field *configschema.Field) ([]byte, error) {
+func renderTable(tmpl *template.Template, field *Field) ([]byte, error) {
 	buf := &bytes.Buffer{}
 	err := executeTableTemplate(tmpl, field, buf)
 	if err != nil {
@@ -33,7 +30,7 @@ func renderTable(tmpl *template.Template, field *configschema.Field) ([]byte, er
 	return buf.Bytes(), nil
 }
 
-func executeTableTemplate(tmpl *template.Template, field *configschema.Field, buf *bytes.Buffer) error {
+func executeTableTemplate(tmpl *template.Template, field *Field, buf *bytes.Buffer) error {
 	err := tmpl.Execute(buf, field)
 	if err != nil {
 		return err
@@ -55,7 +52,7 @@ const durationBlock = "### time-Duration \n" +
 	"each with a unit suffix, such as `300ms`, `-1.5h`, " +
 	"or `2h45m`. Valid time units are `ns`, `us`, `ms`, `s`, `m`, `h`."
 
-func hasTimeDuration(f *configschema.Field) bool {
+func hasTimeDuration(f *Field) bool {
 	if f.Type == "time.Duration" {
 		return true
 	}
